🛠️ refactor : 게임스타트 코드 분리
게임 스타트 코드 분리중입니다.
