+++ src/elements/Button.jsx
@@ -4,6 +4,7 @@ import styled, { css } from 'styled-components';
 const Button = ({ children, ...props }) => {
   return (
     <>
+      {/* 투표때 필요해서 넣었음.. */}
       {children === '투표준비' ? (
         <StyledButton
           type={props.type}
@@ -24,14 +25,6 @@ const Button = ({ children, ...props }) => {
           {children}
         </StyledButton>
       )}
-      {/* <StyledButton
-        type={props.type}
-        onClick={props.onClick}
-        addStyle={props.addStyle}
-        doong={props.doong}
-      >
-        {children}
-      </StyledButton> */}
     </>
   );
 };
